Add taint feature to auto replace tainted node
When a node is tainted by a key, operator will
mark pods on the node as NodeTaintDetected.
If a pods stay in  NodeTaintDetected for long enough, operator will
mark the pod as unhealthy and replace it with replacement logic.

SRE can configure the tainted keys operator should react to.
Each tainted key has a tainted duration which defines how long the pod
should stay in tainted mode before it is marked for replacement.

Change sets include
 * Set logger to output to stdout for replacements_test;
 * Add k8s node concept to unit test framework so that we can
   test node related features (e.g., taint a node, fail a node);
 * Update node and cluster config to propogate its effect to all reconcilation loop;
 * Update NodeTaintDetected condition time with node's tainted time; so that
   a pod is removed "closely" after configured duration after its tainted time;
 * Add taint tests in update_status_test.
   This test focuses on the testing the taint logic isolately;
 * Add taint test in replace_failed_process_groups_test.
   This test focuses on testing the interaction between taint function and other reconciliation loops