Fix an error handling bug in tryConnectionOptions
If none of the connection strings work, it would previously return the
last known connection string and no error. In combination with another
problem, this has led to the operator rotating all coordinators without keeping up
with the latest connection string and then publishing a connection string to
clients that no longer works.