Fix passing incompatible parameters to fdbbackup and fdbrestore via CustomParameters (#2348)
CustomParameters in FoundationDBBackup and FoundationdbRestore CRDs are meant to be passed to backup agents. However, they are also passed to fdbbackup and fdbrestore commands when admincli object is built. An incompatible backup_agent parameter will be rejected by fdbbackup/restore, causing the fdbbackup to never reconcile. We filter out the parameters that are exclusive to backup_agents.